refactor: move DNSTap wire decoding logic to dedicated dnstap_decoder module
